requirements.txt
psycopg2-binary>=2.9.9
qdrant-client>=1.7.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0